Attempt at making the host code more generic and shareable between models. But doesn't work.
