+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "paxos-wasm/shared/utils",
     "paxos-wasm/shared/grpc",
     "paxos-wasm/shared/proto",
+    "paxos-wasm/shared/bindings-types",
 
     "pocs/tcp-server-polling-test",
     "pocs/tcp-server-polling-test/server",
