+++ paxos-wasm/composed-grpc/paxos-coordinator/Cargo.toml
@@ -8,3 +8,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 wit-bindgen = "0.41"
+
